+++ src/main/java/com/github/zastrixarundell/toramsensei/Values.java
@@ -1,9 +1,12 @@
 package com.github.zastrixarundell.toramsensei;
 
 import org.discordbots.api.client.DiscordBotListAPI;
-import org.joda.time.DateTime;
+import redis.clients.jedis.Jedis;
+import redis.clients.jedis.JedisPool;
+import redis.clients.jedis.JedisPoolConfig;
+import redis.clients.jedis.Protocol;
 
-import java.util.List;
+import java.net.URI;
 import java.util.Objects;
 import java.util.Properties;
 
@@ -16,8 +19,6 @@ public class Values
 
     private static int userCount, guildCount;
 
-    public static final String profileImageURL = "https://raw.githubusercontent.com/ZastrixArundell/Toram-sensei/master/images/profile.png";
-
     public final static String toramLogo = "https://toramonline.com/index.php?media/toram-online-logo.50/full&d=1463410056";
 
     public final static String inviteLink = "https://discordapp.com/oauth2/authorize?client_id=600302983305101323&scope=bot&permissions=0";
@@ -28,19 +29,46 @@ public class Values
 
     public final static String donationLogo = "https://raw.githubusercontent.com/ZastrixArundell/Toram-sensei/master/images/patreon.png";
 
-    private static DateTime lastDyeUpdate;
-
-    private static List<String> dyeImages = null;
-
     private static String prefix = ">";
 
     public static void setPrefix(String prefix) { Values.prefix = prefix; }
 
     public static String getPrefix() { return prefix; }
 
-    public static List<String> getDyeImages() { return dyeImages; }
+    private final static JedisPool jedisPool = buildJedisPool();
 
-    public static void setDyeImages(List<String> dyeImages) { Values.dyeImages = dyeImages; }
+    private static JedisPool buildJedisPool() {
+        final JedisPoolConfig config = new JedisPoolConfig();
+        config.setMaxTotal(10);
+        config.setMaxIdle(5);
+        config.setMinIdle(1);
+        config.setTestOnBorrow(true);
+        config.setTestOnReturn(true);
+        config.setTestWhileIdle(true);
+
+        final URI redisURI = URI.create(System.getenv("REDIS_URL"));
+
+        String password;
+        if ((password = redisURI.getUserInfo()) != null)
+            password = password.split(":")[1];
+
+
+        return password == null ?
+                new JedisPool(config, redisURI.getHost(), redisURI.getPort(), Protocol.DEFAULT_TIMEOUT) :
+                new JedisPool(config, redisURI.getHost(), redisURI.getPort(), Protocol.DEFAULT_TIMEOUT, password);
+    }
+
+    public static Jedis getJedis() {
+        Jedis jedis = jedisPool.getResource();
+
+        jedis.clientGetname();
+
+        return jedis;
+    }
+
+    public static void closePool() {
+        jedisPool.close();
+    }
 
     static void getMavenVersion()
     {
@@ -58,9 +86,6 @@ public class Values
 
     public static String getVersion() { return version; }
 
-    public static DateTime getLastDyeUpdate() { return lastDyeUpdate; }
-
-    public static void setLastDyeUpdate(DateTime lastDyeUpdate) { Values.lastDyeUpdate = lastDyeUpdate; }
 
     public static int getUserCount() { return userCount; }