+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### 7.4.0
+- Added feature of automatic dyes updated on `#toram-sensei-dyes`.
+
+### 7.3.0
+- Added Redis usage.
+- Implemented Redis for Upgrade command.
+- Implemented `>month` command.
+
 ### 7.2.0
 - Fixed eveything with `>level` and monster search.
 - Release.